Made the loader a reusable component instead of creating different loaders for every pages that needs a loader.
