+++ client/src/components/pages/Boards/Forms/UpdateForm.jsx
@@ -1,12 +1,12 @@
 import React, { useState, useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { createBoard } from '../../../reducers/BoardReducer';
+import { updateBoardName } from '../../../../reducers/BoardReducer';
 import { useSession } from '@clerk/clerk-react';
 
 import styles from './UpdateForm.module.css';
 
-export default function NewBoardForm({ toggleHandler }) {
-  const [title, setTitle] = useState('');
+export default function UpdateForm({ board, handleToggleUpdateForm }) {
+  const [newTitle, setNewTitle] = useState(board.title);
   const dispatch = useDispatch();
 
   const { session } = useSession();
@@ -15,17 +15,22 @@ export default function NewBoardForm({ toggleHandler }) {
     e.preventDefault();
 
     const formData = {
-      title: title,
+      id: board.id,
+      title: newTitle,
+      userId: board.user_id,
       token: await session.getToken()
     };
 
-    dispatch(createBoard(formData));
+    // dispatch(createBoard(formData));
 
     // console.log(formData);
 
     // Clears the form then close it
-    setTitle('');
-    toggleHandler();
+    // setNewTitle('');
+    // toggleHandler();
+    dispatch(updateBoardName(formData));
+    handleToggleUpdateForm({ ind: null, state: false });
+    console.log(formData);
   }
 
   return (
@@ -34,22 +39,24 @@ export default function NewBoardForm({ toggleHandler }) {
         <form onSubmit={e => onSubmitHandler(e)}>
           <input
             type='text'
-            value={title}
-            name='title'
+            value={newTitle}
+            name='newTitle'
             placeholder='New Board Name'
-            onChange={e => setTitle(e.target.value)}
+            onChange={e => setNewTitle(e.target.value)}
           />
           <div className={styles.formButtons}>
             <input
               type='submit'
-              value='Create'
+              value='Update'
               className={styles.updateBoardButton}
             />
             <input
               type='button'
               value='Cancel'
-              onClick={() => toggleHandler()}
               className={styles.updateBoardButton}
+              onClick={() =>
+                handleToggleUpdateForm({ ind: null, state: false })
+              }
             />
           </div>
         </form>
